Refactor routers: reorganize imports and restore get_client_info endpoint in service_router
